% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_sampler.R
\name{draw_fs}
\alias{draw_fs}
\title{Draw a sample from the feasible set given S and N}
\usage{
draw_fs(i, nspec, nind)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}
}
\value{
a single sample from the feasible set for the SAD
}
\description{
Sample the feasible set.
}
