% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_P_table.R
\name{fill_ks}
\alias{fill_ks}
\title{Build the k table}
\usage{
fill_ks(max_s, max_n)
}
\arguments{
\item{max_s}{the max number of species}

\item{max_n}{the max number of individuals}
}
\value{
table of ks for all combinations of s and n up to max_s and max_n.
}
\description{
Build the table of the maximum abundance for the minimum-abundance species in a community of s species and n total individuals (or the maximum increase in abundance compared to the past species, when building an SAD using a uniform sampler...)
}
