% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ct.R
\name{find_ct}
\alias{find_ct}
\title{Find the central tendency of a collection of RADs.}
\usage{
find_ct(unique_RADs)
}
\arguments{
\item{unique_RADs}{a list of unique RADs with the same S and N}
}
\value{
vector of the central tendency of \code{unique_RADs}.
}
\description{
Find the central tendency following \code{get_hottest_SAD} in Locey & White.
}
