% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ct_criterion.R
\name{new_ct_criterion}
\alias{new_ct_criterion}
\title{New criterion for the likelihood of an RAD}
\usage{
new_ct_criterion(focal_RAD, pool_RADs)
}
\arguments{
\item{focal_RAD}{focal RAD}

\item{pool_RADs}{pool to compare to, must have same S and N as \code{focal_RAD}}
}
\value{
list of log p's and their sum
}
\description{
Testing a new criterion
}
