% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_sampler.R
\name{sample_fs}
\alias{sample_fs}
\title{Draw multiple samples from the feasible set given S and N}
\usage{
sample_fs(nsamples, nspec, nind)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}
}
\value{
matrix of samples
}
\description{
Sample the feasible set.
}
