% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad_sampler.R
\name{send_gnome}
\alias{send_gnome}
\title{Draw a single sample from the feasible set}
\usage{
send_gnome(s, n, ps, ks)
}
\arguments{
\item{s}{the number of species}

\item{n}{the number of individuals}

\item{ps}{the ps table for number of partitions}

\item{ks}{the ks table of max partition sizes}
}
\value{
a vector of length \code{s} whose elements sum up to \code{n}, in non-
decreasing order
}
\description{
Draw a single sample from the feasible set of species abundance
distributions with \code{s} species and \code{n} total individuals, uniformly
}
