% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad_sampler.R
\name{tally_sets}
\alias{tally_sets}
\title{Filter and tally frequency of distinct RADs}
\usage{
tally_sets(sets_matrix)
}
\arguments{
\item{sets_matrix}{matrix of draws from a feasible set (output of \code{sample_fs})}
}
\value{
data frame of unique species abundance vectors (rows are unique vectors, columns are species) and a column \code{set_frequency} of the frequency of each vector in the original set of samples.
}
\description{
Filter a matrix of samples from the feasible set to include only unique vectors, and tally the frequency of those vectors.

The frequency should be approximately equal for all vectors, but for small sample numbers this will not necessarily happen.
}
