% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad_sampler.R
\name{sample_fs}
\alias{sample_fs}
\title{Draw samples from the feasible set given S and N}
\usage{
sample_fs(s, n, nsamples, storeyn = FALSE, storepath = NULL,
  p_table = NULL)
}
\arguments{
\item{s}{the number of species}

\item{n}{the number of individuals}

\item{nsamples}{how many samples to draw}

\item{storeyn}{TRUE/FALSE whether to store the P table.}

\item{storepath}{path to store the P table.}

\item{p_table}{optionally pass P table as an argument}
}
\value{
matrix \code{nsamples} of ranked abundance distributions. Rows are samples and columns are species, sorted from least to most abundant.
}
\description{
Draw samples from the feasible set of species abundance distributions with \code{s} species and \code{n} total individuals
}
